requirements.txt
requests
python-dotenv
playwright
chromadb
